♻ Use modules to compose the build
